quarto dia, arrumado prototipo funcoes

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: davidos- <davidos-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 10:47:04 by davidos-          #+#    #+#             */
-/*   Updated: 2025/10/14 22:22:53 by davidos-         ###   ########.fr       */
+/*   Updated: 2025/10/16 19:40:51 by davidos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*(s + i) != '\0')
